refactor(viewer): extract viewer_schedule.jsx from viewer.jsx
Pure file split, no behavior change. Step 7 of mp-pxxc.
