+++ web-mobile/index.html
@@ -95,6 +95,11 @@
          extracted from viewer.jsx (mp-pxxc step 6).
          Must load after viewer_standings.js and before viewer.js. -->
     <script type="module" src="/dist/viewer_awards.js?v=5"></script>
+    <!-- Schedule + filter components (PlayerMultiFilter, applyFilters, matchHighlightedBy,
+         ScheduleViewer, ViewerSchedule, TWMatch, buildPlayerMatchHighlight,
+         buildWatchlistUpcoming), extracted from viewer.jsx (mp-pxxc step 7).
+         Must load after viewer_match.js and viewer_watchlist_core.js, and before viewer.js. -->
+    <script type="module" src="/dist/viewer_schedule.js?v=5"></script>
     <script type="module" src="/dist/viewer.js?v=5"></script>
     <script type="module" src="/dist/admin_helpers.js?v=5"></script>
     <script type="module" src="/dist/admin_shell.js?v=5"></script>
